cr50: fix ccd's ec_uart_bitbang_en enable
The ec bitbang enable flag is initially stored as a boolean. However, the
symbolic naming logic in servod maps integers to symbolic names. This
causes an issue when on initialization the flag is queried, as the map
does not contain a key for 'False' but rather only for 0. The solution
is to store the flag as an integer from the beginning, and make sure
that all returns through the '_Get_' function are cast to int first.

BUG=b:157789871
(copying from namyoon@):
TEST=sudo servod -b puff --allow-dual-v4
     dut-control ccd_cr50.ec_uart_bitbang_en
ccd_cr50.ec_uart_bitbang_en:off

TEST=sudo servod -b puff --allow-dual-v4
     dut-control servo_type
servo_type:servo_v4_with_ccd_cr50
     dut-control ec_uart_bitbang_en
ec_uart_bitbang_en:off

Change-Id: I4150463739ea9b5edcc0e5c8daba5f5e840dd76b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2225189
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Namyoon Woo <namyoon@chromium.org>