scratch: add marked active
This expands the functionality of the scratch to now allow for instances
to be marked as 'active' as well i.e. they are active to receive servod
traffic. This information can now be used to have a tool that monitors
when an instance is active to run, rather than using generous timeouts
until the instance is responsive.

It also adds necessary new unit-tests to make sure there is no
regression with this functionality.

BUG=b:158341159
TEST=python2 -m pytest scratch_test.py

TEST=sudo servod -b soraka &

Change-Id: I692d5b1ceef5b501c60ffef73bfc11903a637857
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2233881
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>