usb_image_manager: add proper raises section
a method call for getting the usb stick's /dev path now raises an error
if the ports on the internal hub where a usb stick can be found are not
provided in the xml overlay. This change adds the proper documentation
for that.

BUG=None
TEST=pylint

Change-Id: I1d46ae3b3ff37bb1b941791635865e6ba9a41844
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2247231
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>