system_config: introduce tagging
This change introduces the ability to tag controls and store that in the
SystemConfig to make introspection easier.
This first iteration is kept simple, and additional features or
complexity can be added if necessary. Specifically, this means:
- the main control name can be found, none of the aliases
- the tags need to be defined in the get dictionary (relevant for
controls that define get and set)
- tag lookups are explicitly generated to avoid having new controls
overwrite previous tags etc through 'clobber_ok'

This change only introduces the feature, and does not leverage it. The
subsequent change leverages it to simplify the controls to find the INA
controls.

BUG=chromium:823677
TEST=// with subsequent CL
TEST=python2 -m pytest system_config_unittest.py

Change-Id: Ib6f4ecfdeada55198bb149e197de6c59c2781c5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2267863
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>