active_v4_device: check SPI in ccd active check
Right now the active_v4_device check fails when the EC and AP are off.
active_v4_device checks EC and AP uart to determine if ccd is active.
However, cr50 will disable uart when the EC or AP are off. This means
UARTAP and UARTEC won't be in the ccdstate output even though ccd is
active. If you try to get active_v4_device in this state, it'll return
'neither'. EC-EFS2 systems only check APUART, so it'll fail whenever the
AP is off.

This change adds a SPI check to _using_ccd, becaus SPI is always active
when ccd is open and connected. It is not dependent on the rest of the
system state.

BUG=b:158338993
TEST=dut-control active_v4_device

Change-Id: I550175e9355e6f0a84188f4d3d89f7ea50446299
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2274059
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>