servod: Make Servod instance as interface object a weakref proxy.
This breaks a reference cycle for drivers with "servo" interface configured
in their XML data/ file, which is also the default interface if none is
specified.

Why this is safe: There should be no need for driver objects to keep a
servo object alive.

BUG=b:163418494
TEST=dev-util/hdctools builds, servod starts, and
"dut-control active_v4_device" continues to function - it uses "servo"
interface and would break if the weakref was not functional.

Change-Id: If397ea01b55eea765a1c821377b1794ee63d0ac4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2371275
Tested-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Matthew Blecker <matthewb@chromium.org>