stm32usb: release resources
When repeatedly opening a new interface to the same usb device, pyusb
might be slow the release the resources and might fail.
According to the docs, even if the device is used again, this is fine as
the resources get reallocated on request.
https://github.com/pyusb/pyusb/blob/cf88dfcd3b38afd00ec9fc7229213c8e8aa42118/usb/util.py#L221

BUG=b:142143881
TEST=servo_mfg v4p1 // subsequent CLs
TEST=sudo servod -b atlas // v4 + ccd
     unplug ccd, replug ccd
     dut-control ec_board
ec_board:atlas

Change-Id: I8ed92df0a437ece0372aa4a4b1deb0b47baa3659
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2407394
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>