servod: Strip non-ascii chars from uart output
Depending on whether python2 or python3 is used, the
str.encode method may not work properly.
For python2, an error occurs because the codec will initially perform
a decode before the encode, and the decode will fail on a non-ascii
char. For python3, there is no decode method for a string.

The solution is to use a method that works for both python2 and python3.

BUG=b:164965140
TEST=Confirm that firmware_PDConnect passes correctly.

Change-Id: I389991c76fdf6b238c18b3ea93fb004f25a95581
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2439971
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Andrew McRae <amcrae@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>