servod: Add timeout on the pty flush
If the servo is in a loop where it sends output on a continuous basis,
the flush operation needs to eventually timeout. If the servo sends
commands out faster than the timeout provided to the expects() call,
it would never advance. This blocked the servo_updater from advancing.

BUG=b:162628581
TEST=Connect Servo_v4p1 to DUT validate servo_updater works
TEST=Modify Servo firmware to generate many console messages
    in a loop. Validate that the flush operation succeeds and
    servo updater still works.

Tested-by: Brian Nemec <bnemec@chromium.org>
Change-Id: I3100a91b3e3daf43998568e41a1a882b20782ce5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2486867
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Commit-Queue: Brian Nemec <bnemec@chromium.org>