servod: add --recovery mode
This CL adds a basic 'recovery mode' notion. The idea is to build this
out so that different startups that would normally fail can succeed so
that the caller might no introspection and try to recover the issues.
This initial CL simply adds the functionality to 'request' a recovery
mode startup, and the ability for other code to check if that has been
requested, to avoid raising an error but rather log it, and continue.
The first use-case is servo v4 startup with a --board flag. Normally,
when this happens, we fail if no dut-controller (micro, ccd_cr50) can be
found. In recovery mode we continue so that we can use the servo v4
controls to perform introspection on the instance, and potentially
recover the ccd_cr50 device.

BUG=b:159755652, b:173059305

TEST=sudo servod -b grunt // no issue
// observe normal startup

TEST=sudo servod -b grunt // no cr50 enumerated
// observe failure to start up and voltage logs

TEST=sudo servod -b grunt --recovery_mode // no cr50 enumerated
// observe normal startup with voltage logs and warning
2020-11-13 18:11:42,020 - ServoV4PostInit - INFO - Will continue startup
as recovery mode has been requested

Change-Id: I8a0cd47476c79ca2748b87128c1f85624fa5142e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2538580
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>