servod: add fault and sbu voltage low
This change adds a basic fault driver that acts like a software flag
i.e. it can be 'on' or 'off', with default 'off'. It can just be set.
This is then used to create a control named
`dut_sbu_voltage_float_fault` that is 'on' if the SBU voltage (SBU1 or
SBU2) is between threshold values i.e. floating. This, in combination
with recovery mode, allows the caller to see if the DUT's ccd_cr50 USB
device is not enumerating because the SBU signal is not being forwarded
correctly.

BUG=b:159755652, b:173059305

// showing functionality of the software flag
TEST=sudo servod -b volteer // normal start-up
     dut-control dut_sbu_voltage_float_fault
dut_sbu_voltage_float_fault:off
     dut-control dut_sbu_voltage_float_fault:on
     dut-control dut_sbu_voltage_float_fault
dut_sbu_voltage_float_fault:on
     dut-control dut_sbu_voltage_float_fault:off
     dut-control dut_sbu_voltage_float_fault
dut_sbu_voltage_float_fault:off

// showing floating voltage detection (grunt)
TEST=sudo servod -b grunt --recovery-mode// bad startup
     dut-control dut_sbu_voltage_float_fault
dut_sbu_voltage_float_fault:on

Change-Id: If20c5a937577084ba646afe5259c4496f422f47b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2538581
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>